divide popup and photos
